requirements.txt
fastapi
pydantic
uvicorn
google-generativeai
serpapi
python-dotenv
streamlit
speechrecognition
pygame
gtts
requests
